% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEcompare.R
\name{DEcompare}
\alias{DEcompare}
\title{DEcompare}
\usage{
DEcompare(deg_list, threshold = 1.5, genes = NULL, xlim = c(-10, 10), ylim = c(-10, 10), xlab = "Contrast1", ylab = "Contrast2", main = "Comparison of Log2FC", subtitle = paste(xlab, "vs", ylab), color_corners = c("pink", "lightgreen", "cornflowerblue", "yellow"), alpha_corners = c(.7))
}
\arguments{
\item{deg_list}{List of lenght 2. Must contain data frames with the columns Geneid, log2FoldChange, padj and DEG. Values in 'Geneid' must be the same in both data frames.+}

\item{threshold}{Numeric of lenght 1. Threshold for the log2FoldChange. It is used to draw horizontal and vertical lines. Default: 1.5.}

\item{genes}{Charachter of indefined lenght or NULL. Genes to be written in the plot. Default: NULL.}

\item{xlim}{Numeric of length 2. The limits of the Y axis. Default: c(-10, 10)}

\item{xlab}{Character of lenght 1. Name of the X axis. Default: "Contrast1"}

\item{ylab}{Character of lenght 1. Name of the Y axis. Default: "Contrast2"}

\item{subtitle}{Character of lenght 1. Subtitle of the plot. Default: paste(xlab, "vs", ylab)}

\item{alpha_corners}{Numeric of length 1 to 4. If too short, it will be recycled. Alpha (transparency) value for the vars. Default: 0.7}

\item{title}{Character of lenght 1. Title of the plot. Default: "Comparison of Log2FC"}

\item{colors}{Character of length 1 to 4. If too short, it will be recycled. Colors for each of the corners: top-left, bottom-left, top-right, bottom-right. Default: c("pink", "lightgreen", "cornflowerblue", "yellow")}
}
\description{
DEcompare
}
\author{
amitjavilaventura
}
