% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionCor.R
\name{expressionCor}
\alias{expressionCor}
\title{expressionCor}
\usage{
expressionCor(
  df,
  genes = NULL,
  samples = NULL,
  samples_order = samples,
  corr_method = "pearson",
  plot_type = "lower",
  plot_diagonal = TRUE,
  plot_size = 85,
  plot_border = TRUE,
  plot_colors = c("Gold3", "White", "Cornflowerblue"),
  plot_title = NULL,
  plot_subtitle = NULL,
  plot_caption = NULL,
  cell_border = "Black",
  legend_pos = "right",
  legend_size = c(8, plot_size),
  legend_limits = c(-1, 1),
  legend_breaks_by = 0.5,
  legend_title = paste(stringr::str_to_sentence(corr_method), "correlation", sep = " "),
  legend_title_size = 10,
  legend_frame_width = 0.5,
  ticks = F,
  ticks_color = "Black",
  coeffs_color = "Black",
  coeffs_size = 4,
  title_hjust = 0.5,
  title_face = "plain",
  title_size = 12,
  subtitle_face = "italic",
  subtitle_size = 11,
  caption_size = 6,
  axis_text_size = 8,
  axis_text_color = "black",
  by_groups = F
)
}
\arguments{
\item{df}{Dataframe with a 'Geneid' column and several columns with numerical expression data for different samples, such as TPMs or Log2FC.}

\item{genes}{Character or NULL. Names of the genes to be used for the correlation analysis. They must present in the column 'Geneid' of 'df'. If NULL, all genes in the data frame are used. Default: NULL}

\item{samples}{Character or NULL. Names of the samples to be used for the correlation analysis. They must be column names of 'df'. If NULL, all samples are used. Default: NULL.}

\item{corr_method}{Character of length 1. Correlation method to be passed through \code{cor()}. One of \code{"pearson"}, \code{"spearman"} or \code{"kendall"}. Default: "pearson".}

\item{plot_type}{Character of length 1. Type of the correlation plot. One of \code{"full"}, \code{"upper"} or \code{"lower"}. Default: "lower".}

\item{plot_diagonal}{Logical of length 1. Whether to plot the diagonal or not if the plot type is upper or lower. Default: TRUE.}

\item{plot_size}{Numerical of length 1. Size in milimeters of the plot. The same number is used for plot height and plot width Default: 85.}

\item{plot_border}{Logical of length 1. Whether to draw a border in the panel or not. Default: TRUE.}

\item{plot_colors}{Character of length 3. Colors of the lower, midpoint and higher limits of the scale. Default: c("Gold3", "White", "Cornflowerblue").}

\item{plot_title}{Character of length 1 or NULL. Title of the plot. Defalut: NULL.}

\item{plot_subtitle}{Character of length 1 or NULL. Subtitle of the plot. Defalut: NULL.}

\item{plot_caption}{Character of length 1 or NULL. Caption of the plot. Defalut: NULL.}

\item{cell_border}{Charachter of length 1, NULL or NA. Color of the cell border. Default: "Black".}

\item{legend_pos}{Charachter of length 1. Position of the legend, to be passed through \code{ggpubr::theme_pubr()}. Default: "right".}

\item{legend_size}{Numerical of length 2. Size of the legend. The first element is the width and the second is the height of the legend bar. Default: c(8,plot_size)}

\item{legend_limits}{Numerical of length 2. Lower and upper limits of the correlation scale. Default: c(-1,1).}

\item{legend_breaks_by}{Numerical of length 1. Size of the breaks in the legend. Default: 0.5.}

\item{legend_title}{Character of length 1 or NULL. Title of the legend. Default: \code{paste(stringr::str_to_sentence(corr_method), "correlation", sep = " ")}.}

\item{legend_title_size}{Numerical of length 1. Size of the legend title. Default: 10.}

\item{legend_frame_width}{Numerical of length 1. Width of the line surrounding the colorbar legend. Default: 0.5.}

\item{ticks}{Logical of length 1. Whether to show the ticks of the colourbar legend or not. Default: FALSE.}

\item{ticks_color}{Character of length 1. Color of the ticks of the colourbar legend. Default: "Black".}

\item{coeffs_color}{Character of length 1 or NULL. Color of the correlation coeficients to be plotted. If NULL, no correlation coefficients are plotted. Default: "Black".}

\item{coeffs_size}{Numerical of length 1. Size of the correlation coefficients. Default: 4.}

\item{title_hjust}{Numerical of length 1. Horizontal justification of the title and the subtitle. Default: 0.5.}

\item{title_face}{Character of length 1. Face of the title text. One of "plain", "italic", "bold". Default: "plain".}

\item{title_size}{Numerical of length 1. Size of the title text. Default: 12.}

\item{subtitle_face}{Character of length 1. Face of the subtitle text. One of "plain", "italic", "bold". Default: "italic".}

\item{subtitle_size}{Numerical of length 1. Size of the subtitle text. Default: 11.}

\item{caption_size}{Numerical of length 1. Size of the plot caption. Default: 6.}

\item{axis_text_size}{Numerical of length 1. Size of the text in the axes. Default: 7.}

\item{axis_text_color}{Character of length 1. Color of the text in the axes. Default: "Black".}

\item{by_groups}{Logical of length 1. If TRUE, the function searches an extra column named 'group', whose values will be used to separate the plot into facets using facet_wrap(). Default: FALSE.}
}
\description{
Function that takes a data frame with a Geneid column and several columns for
expression data of different samples (e.g. TPMs, Log2FC...) and draws a correlation plot
}
\author{
amitjavilaventura
}
