% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVennCounts.R
\name{getVennCounts}
\alias{getVennCounts}
\title{getVennCounts}
\usage{
getVennCounts(peaks, conds = names(peaks), conds_order = conds, plot = F)
}
\arguments{
\item{peaks}{List of dataframes with the genomic coordinates of the regions to overlap. Dataframes must contain the columns seqnames, start, end.}

\item{conds}{Character with the same length as the 'peaks' list. The names that are given to the diferente objects in the 'peaks' list. Default: 'names(peaks)'}

\item{conds_order}{Character of the same length as 'conds'. Same values as in 'conds' but with the desired order of priority. Default: 'conds'.}

\item{plot}{Logical of lenght 1. Whether to draw the the 'makeVennDiagram()' plot or not. Default: FALSE}
}
\description{
getVennCounts
}
\seealso{
\code{ChIPpeakAnno::makeVennDiagram}
}
\author{
amitjavilaventura

Function that calls 'ChIPpeakAnno::makeVennDiagram()' to obtain the venn counts.
It also does a list of peaks and with several variables corresponding to the presence of these peaks in each set.
It's a helper for other functions to draw Venn diagrams or UpSet plots for peak intersections.
}
