% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggUpsetPeaks.R
\name{ggUpsetPeaks}
\alias{ggUpsetPeaks}
\title{ggUpsetPeaks}
\usage{
ggUpsetPeaks(
  peaks,
  conds = names(peaks),
  conds_order = conds,
  order_by_freq = T,
  num_size = 4,
  title = NULL,
  subtitle = NULL,
  caption = NULL
)
}
\arguments{
\item{peaks}{List of dataframes with the genomic coordinates of the regions to overlap. Dataframes must contain the columns seqnames, start, end.}

\item{conds}{Character with the same length as the 'peaks' list. The names that are given to the diferente objects in the 'peaks' list. Default: 'names(peaks)'}

\item{conds_order}{Character of the same length as 'conds'. Same values as in 'conds' but with the desired order of priority. Default: 'conds'.}

\item{order_by_freq}{Logical of length 1. If TRUE (the default), the intersections are arranged by descending order of size. Default: TRUE.}

\item{num_size}{Numeric of length 1. Size of the labels indicating the size of each intersection. Default: 4.}

\item{title}{Character of length 1. Title of the plot. Default: NULL.}

\item{subtitle}{Character of length 1. Subtitle of the plot. Default: NULL.}

\item{caption}{Character of length 1. Caption written at the bottom of the plot. Default: NULL.}
}
\description{
ggUpsetPeaks
}
\seealso{
plyranges

Function that calls 'getVennCounts' and draws an ggplot2-based upset plot.

ggUpsetPeaks(peaks, conds = names(peaks), conds_order = cond, order_by_freq = T, num_size = 4, title = NULL, subtitle = NULL, caption = NULL
}
\author{
amitjavilaventura
}
