% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVennBed.R
\name{ggVennBed}
\alias{ggVennBed}
\title{ggVennBed}
\usage{
ggVennBed(
  a,
  b,
  setnames = c("A", "B"),
  stranded = T,
  color = c("blue", "gold3"),
  alpha = 0.5,
  namesize = 7,
  labsize = 5,
  title = NULL,
  subtitle = NULL,
  title_size = 12,
  subtitle_size = 11,
  title_face = "plain",
  subtitle_face = "italic"
)
}
\arguments{
\item{a}{Data frame or character of length 1. Path to a BED file or data frame with BED format. If data frame, it will be written to a temporary file.}

\item{b}{Data frame or character of length 1. Path to a BED file or data frame with BED format. If data frame, it will be written to a temporary file.}

\item{setnames}{Character of length 2. Names of the data sets of 'a' and 'b'. Default: c("A", "B").}

\item{stranded}{Logical of length 1. If TRUE, the intersections are done regarding the strand. Default = TRUE.}

\item{color}{Character of length 2. Color of the Venn circles and the labels indicating the number of regions. Default: c("blue", "gold3").}

\item{alpha}{Numeric of length 1. Transparency of the circles. Default: 0.5.}

\item{namesize}{Numeric of length 1. Size of the names of each set. Default: 7.}

\item{labsize}{Numeric of length 1. Size of the labels indicating the number of unique and overlapping regions.}
}
\description{
Does the intersection of two BED files (A and B) and
draws the intersection with the unique peaks of A and B and
the intersections of A with B and B with A.
}
\seealso{
\code{bedtoolsr::bt.intersect}

\code{ggvenn::ggvenn}
}
\author{
amitjavilaventura
}
