% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDendogram.R
\name{plotDendogram}
\alias{plotDendogram}
\title{plotDendogram}
\usage{
plotDendogram(hclust_data, axis = "rows")
}
\arguments{
\item{hclust_data}{Output from \code{hclust()} function.}

\item{axis}{Characteer of length 1. Either "rows" or "cols". If rows, the dendogram is made to be placed in the left of the Y axis; if cols, the dendogram is made to be placed in the top of the X axis.}
}
\description{
Helper function that takes clustering data (hclust()) as input and
plots a dendogram using ggplot and ggdendro
}
\seealso{
\code{ggpubr}

\code{ggdendro}
}
\author{
amitjavilaventura
}
