% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{theme_barDEGs}
\alias{theme_barDEGs}
\title{theme_barDEGs}
\usage{
theme_barDEGs(xaxis = F, yaxis = F)
}
\arguments{
\item{xaxis}{Logical of length 1. Whether to draw the text of the X axis (number of DE) or not. Default: FALSE.}

\item{yaxis}{Logical of length 1. Whether to draw the text of the Y axis (contrast names) or not. Default: FALSE.}
}
\description{
theme_barDEGs
}
\author{
amitjavilaventura

Helper for the 'barDEGs' function.
}
