% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{theme_chromReads}
\alias{theme_chromReads}
\title{theme_chromReads}
\usage{
theme_chromReads(main.size = 11, sub.size = 10, axis.size = 9)
}
\arguments{
\item{xaxis}{Logical of length 1. Whether to draw the text of the X axis (number of DE) or not. Default: FALSE.}

\item{yaxis}{Logical of length 1. Whether to draw the text of the Y axis (contrast names) or not. Default: FALSE.}
}
\description{
theme_chromReads
}
\author{
amitjavilaventura

Helper for the 'chromReads' function.
}
