% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsetPeakOverlap.R
\name{upsetPeaks}
\alias{upsetPeaks}
\title{upsetPeaks}
\usage{
upsetPeaks(peaks, names = names(peaks), names_order = names)
}
\arguments{
\item{peaks}{List of dataframes with the genomic coordinates of the regions to overlap. Dataframes must contain the columns seqnames, start, end.}

\item{conds}{Character with the same length as the 'peaks' list. The names that are given to the diferente objects in the 'peaks' list. Default: 'names(peaks)'}

\item{conds_order}{Character of the same length as 'conds'. Same values as in 'conds' but with the desired order of priority. Default: 'conds'.}

\item{order.by}{Same as in UpSetR::upset(). One of "freq", "degree" or both.}

\item{mainbar.y.label}{Same as in UpSetR::upset().}

\item{sets.x.label}{Same as in UpSetR::upset().}

\item{...}{Further arguments to be passed through 'UpSetR::upset()'.}

\item{title}{(Not available yet) Character of length 1 or NULL. Title of the plot shown at the top. Passed through grid.arrange(). If NULL, title is not shown. Default: NULL}
}
\description{
upsetPeaks
}
\seealso{
plyranges

UpSetR
}
\author{
amitjavilaventura
}
