% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromRegions.R
\name{chromRegions}
\alias{chromRegions}
\title{chromRegions}
\usage{
chromRegions(chrom_sizes, regions_list, regions_names = names(regions_list), regions_order = names(regions_list), chr_exclude = c("Un", "Random", "JH", "GL", "\\."), chr_order = c(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,"X","Y","M","MT"), col_by_strand = FALSE, regions_names = names(regions_list), regions_colors = sample(x = colors(), size = length(regions_list), replace = F), title = NULL, subtitle = NULL, caption = NULL, xlab = "Position (bp)", ylab = "Chromosome", legend = "right", draw_points = TRUE, coord_flip = TRUE)
}
\arguments{
\item{chrom_sizes}{Character of lenght 1 or dataframe. Path to the chrom.sizes file or data frame with the names of the chromosomes and their sizes. The file must not contain column names.}

\item{regions_list}{(Named) List of character vectors or dataframes. List with paths to BED/TSV files or dataframes with chromosome names, start, end, region id, length and strand. Strand must be "+", "-" or ".".}

\item{regions_names}{Character of length equal to regions_list. Names of the input regions. Default: names(regions_list).}

\item{regions_order}{Character of length equal to regions_list. Names of the input regions in the desired order. Default: names(regions_list).}

\item{colors}{Character of length 2 or of length equal to regions_list. Colors of the input regions. If col_by_strand is FALSE, the length must be equal to the input regions, if col_by_strand is TRUE, the length must be 2. Default: sample(x = colors(), size = length(regions_list), replace = F).}

\item{color_by}{Character of length 1. One of 'region', 'strand' or 'extra'. If 'region', colours by the regions in 'regions_list'. If 'strand', colours by strand. If 'extra' colours by the grouping provided field in 'extra_info' (if provided).}

\item{chr_exclude}{Character. Regular expressions to match the chromosomes names to exclude . Default: c("Un", "Random", "JH", "GL", "\\.").}

\item{chr_order}{Character. Chromosome names written in the order to be plotted. The names must matchthe chromosome names in chrom_sizes and regions. Default: c(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,"X","Y","M","MT")}

\item{title}{Character of length 1 or NULL. Title of the plot. Default: NULL}

\item{subtitle}{Charachter of length 1 or NULL. Subtitle of the plot. Default: NULL.}

\item{caption}{Character of length 1, TRUE or NULL. Caption to be written in the bottom-right corner. If TRUE, it will write the number of regions in the input; if charachter, it will write the written caption. Default: NULL.}

\item{ylab}{Character of length 1. Title of the X axis. Default: ""}

\item{xlab}{Character of lenght 1. Title of the Y axis. Default: "Chromosome".}

\item{legend}{Character of length 1. Position of the legend, passed through ggpubr::theme_pubr(). One of c("top", "bottom", "left", "right", "none"). Default: "right".}

\item{draw_points}{Logical of lenght 1. If TRUE, draws points in each of the drawn regions. Default: TRUE}

\item{coord_flip}{Logical of lenght 1. If TRUE, change the position of the axes (Y axis is position and X is chromosome). Default: FALSE}

\item{extra_info}{NULL or (named) list of length equal to 'regions_list', and with the same order. List of path to files or list of data.frames with two columns containing the id column of the regions contained in each set of 'regions_list' and a column with extra (discrete) information. Default: NULL.}

\item{size_y_text}{NULL or numeric of length 1. If NULL, do not draw the text on the Y axis (position in bp). If numeric, the size of the text in the Y axis. Default: NULL}
}
\description{
Function that draws the chromosomes and the position of the desired regions.
}
\author{
amitjavilaventura
}
