% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVennPeaks.R
\name{ggVennPeaks}
\alias{ggVennPeaks}
\title{ggVennPeaks}
\usage{
ggVennPeaks(peak_list, peak_names = names(peak_list), percent = T, fill = c("blue", "gold3"), alpha = .4, color = "black", text_color = "black", name_size = 5, label_size = 3, title = "", subtitle = "")
}
\arguments{
\item{peak_list}{List of dataframes with the genomic coordinates of the regions to overlap. Dataframes must contain the columns seqnames, start, end.}

\item{peak_names}{Character with the same length as the 'peaks' list. The names that are given to the diferente objects in the 'peaks' list. Default: 'names(peaks)'}

\item{percent}{Logical of length 1. Whether to show the percentage corresponding to each part of the intersection. Default = T.}

\item{stranded}{Logical of length 1. If TRUE, it considers the strand information to do the overlaps. Default = FALSE.}

\item{in_fill}{Character of length equal to 'peak_list'. Colors of each set in 'peak_list'. Default: c("blue", "gold3").}

\item{alpha}{Numeric of length 1. Transparency of the the Venn circles. Default: 0.4}

\item{out_color}{Character of length 1 or equal to 'peak_list'. Color of the circles outer lines. Default: "black"}

\item{name_color}{Character of length 1 or equal to 'peak_list'. Color of the titles of each set of peaks. Default: "black"}

\item{text_color}{Character of length 1. Color of the text inside the Venn circles (i.e. number of peaks and percentage in each intersection). Default: "black"}

\item{name_size}{Numeric of length 1. Size of the name of each peak set. Default: 5}

\item{label_size}{Numeric of length 1. Size of the text inside the Venn circles. Default: 3}

\item{title}{Character of length 1. Title of the plot. Can be NULL. Default: ""}

\item{subtitle}{Character of length 1. Title of the plot. Can be NULL. Default: ""}
}
\description{
Function that calls 'getVennCounts()' and draws a VennDiagram of
peak intersections using the \code{ggvenn} package.
}
\seealso{
\code{ggvenn::ggvenn}

\code{ChIPseeker::makeVennDiagram}
}
\author{
amitjavilaventura
}
